requirements.txt
openai==1.54.4
python-dotenv==1.0.1
pyttsx3==2.90
mutagen==1.47.0
pydub==0.25.1
librosa==0.10.1
SpeechRecognition==3.10.0